cambio img

+++ src/main/java/com/john/backend_gestion_restaurantes/servicios/imagenes/FirebaseStorageService.java
@@ -65,15 +65,20 @@ public class FirebaseStorageService {
     public String deleteFile(String publicUrl) throws IOException{
         // Obtener el nombre del blob de la URL pública
         String blobName = publicUrl.substring(publicUrl.lastIndexOf('/') + 1, publicUrl.indexOf("?"));
-        // Eliminar el blob del Firebase Storage
-        Bucket bucket = storageClient.bucket(bucketName);
-        Blob blob = bucket.get(blobName);
-        if (blob != null) {
-            blob.delete();
-            return "ok";
-        } else {
-            throw new IllegalArgumentException("La imagen no existe en Firebase Storage");
+        if(blobName != "b64b193d-3cc3-4ee9-ae33-b2033dbdceb9.jpeg" && blobName != "ee563e42-e7c3-4734-8077-b99ad71dc145.jpeg"){
+             // Eliminar el blob del Firebase Storage
+            Bucket bucket = storageClient.bucket(bucketName);
+            Blob blob = bucket.get(blobName);
+            if (blob != null) {
+                blob.delete();
+                return "ok";
+            } else {
+                throw new IllegalArgumentException("La imagen no existe en Firebase Storage");
+            }
+        }else{
+            return "0k";
         }
+       
     }
 
     public String updateFile(String publicUrl, String newFile) throws IOException {
